on-prem/requirements.txt
faker
psycopg2-binary

